dust functionality

+++ src/main/java/com/ruinscraft/powder/objects/Dust.java
@@ -5,12 +5,12 @@ import org.bukkit.Particle;
 public class Dust {
 	
 	private Particle particle;
-	private float radius;
-	private float height;
+	private double radius;
+	private double height;
 	
 	private long frequency;
 	
-	public Dust(Particle particle, float radius, float height, long frequency) {
+	public Dust(Particle particle, double radius, double height, long frequency) {
 		this.particle = particle;
 		this.radius = radius;
 		this.height = height;
@@ -21,11 +21,11 @@ public class Dust {
 		return particle;
 	}
 	
-	public Float getRadius() {
+	public Double getRadius() {
 		return radius;
 	}
 	
-	public Float getHeight() {
+	public Double getHeight() {
 		return height;
 	}
